server/requirements.txt
Django
djangorestframework
markdown       # Markdown support for the browsable API.
django-filter  # Filtering support
drf-nested-routers

